requirements.txt
fastapi
uvicorn
Pillow
python-multipart
websockets
requests
opencv-python-headless
huggingface_hub
